Associate talks with their proposer.
It's reasonable to want to know who proposed a talk.

We've placed this assocation on the User, but tried to make it
transparently available via account. Basically, we're trying to hide the
existence of the User instance as it's an implementation detail to hold
our domain-specific information without muddling it with Devise.