Allow proposal descriptions to have some formatting.
Otherwise they end up as just one long line, and that's not very useful for communicating points. I like Markdown, so that's what we have.

I've refactored some of the test helper methods to make this test a bit less brittle (the definition of the container for a proposal is still a single method, for example).