Idea: proposal on the left, feedback on the right.
The point here is that the proposal is important, and so when reading feedback it should always be done with the proposal to hand. I guess what I *actually* want is for the proposal to somehow always be visible when the feedback is scrolled though; I can think of a few implementations, but they're all a bit weird, and I'm not 100% sure of that anyway.

But regardless, I think it's good to emphasis the proposal when reading suggestions; if the suggestions end up dominating the page, then they will surely end up becoming too self-referential.