Make this test more deterministic.
Previously this was failing with certain seeds (such as 33176) when running as part of the full suite. Debugging revealed this was because the test was assuming that the user's name didn't end in an s, so that the possessive form matched the expectation in the test.

The simplest fix is to hard-code the name; possessive functionality can be tested elsewhere if it isn't already.