Update scores when proposals and suggestions are made.
This also revealed that I'd misunderstood what proposals of interest meant; I'd assumed it was a proposal of yours that someone had commented on, instead it's just the proposals that you've commented on.

This means that the scores are currently wrong; however, I'm going to try to add a scoring in to match my original intention.