Get the build green again
Our markdown method returns '' instead of nil.  It breaks the transparency of the method (e.g. pass in nil, get nil back), but it makes it safer when used in our rss feeds, as apparently builder doesn't like cdata!-ing nil.

On the selection page we might be showing more than one instance of a proposal (your selections + top 10), so we need to prefix the ids to disambiguate them.