Put a markdown hint on all the markdownable inputs
Use a custom markdown input and have it augment the label text.  In passing we also turn application.css into a scss file and use @import instead of require_tree . as this seems to be the only way to share variables from bootstrap for use in our own scss files.