Provide an .in_modification_order scope for sorting proposals
Looks at updated_at and the updated_at on any suggestions, taking the largest available.  Does a sneaky virtual attribute load on it too so that #last_modified won't try to calculate it.

Use it on the dashboard to order proposals by modification date so that the "you should look at..." sections maybe make a bit more sense.