Implement a very simple feature flag
This allows people to switch a feature off or on with a config var.

This is different (and, I hope, complementary) to the mode approach.

My usecase is that UIKonf doesn't want to use the Motiviations feature,
so we want to switch it off, preferably without messing with any code.

Motivations may have several interactions with modes - Ruby Manor wants
to know your motivations before the event, but probably not in Archive
mode. The 'No motivations at all' need is orthogonal to the 'Motivations
then but not now' need.

I've deliberately kept this implementation as light as possible.